requirements.txt
# TrendXL Requirements
# Compatible with Python 3.8-3.13

fastapi==0.110.0
uvicorn[standard]==0.29.0
ensembledata==0.2.6
openai==1.30.0
requests==2.31.0
pydantic==2.8.2
python-dotenv==1.0.0
seatable-api==2.6.0
httpx==0.26.0
python-multipart==0.0.7

# Additional dependencies for better compatibility
pydantic-settings>=2.0.0
email-validator>=2.0.0
